- added new method 'GetAllBooks()' - fixed unit tests

+++ BookLibraryManagerTests/XmlBookListSaverTests.cs
@@ -8,18 +8,16 @@ namespace BookLibraryManagerTests;
 public class XmlBookListSaverTests
 {
     [Fact]
-    public void SaveLibrary_ShouldSaveLibraryToFile()
+    public void SaveLibrary_ValidLibrary_Success()
     {
-
         // Arrange
         var mockLibrary = new Mock<ILibrary>();
         mockLibrary.SetupGet(lib => lib.Id).Returns(1);
-        mockLibrary.SetupGet(lib => lib.BookList).Returns(new List<Book>
-        {
+        mockLibrary.SetupGet(lib => lib.BookList).Returns(
+        [
             new() { Id = 1, Author = "Author1", Title = "Title1", PageNumber = 1 },
             new() { Id = 2, Author = "Author2", Title = "Title2", PageNumber = 2 }
-        });
-
+        ]);
         var saver = new XmlBookListSaver();
         var pathToFile = "testLibrary.xml";
 
@@ -30,7 +28,31 @@ public class XmlBookListSaverTests
         Xunit.Assert.True(result);
         Xunit.Assert.True(File.Exists(pathToFile));
 
-        // Clean up
-        File.Delete(pathToFile);
+        // Cleanup
+        if (File.Exists(pathToFile))
+        {
+            File.Delete(pathToFile);
+        }
+    }
+
+    [Fact]
+    public void SaveLibrary_NullLibrary_ThrowsException()
+    {
+        // Arrange
+        var xmlBookListSaver = new XmlBookListSaver();
+        var pathToFile = "testLibrary.xml";
+
+        // Act & Assert
+        bool result;
+        try
+        {
+            Xunit.Assert.Throws<InvalidCastException>(() => xmlBookListSaver.SaveLibrary(null, pathToFile));
+            result = true;
+        }
+        catch
+        {
+            result = false;
+        }
+        Xunit.Assert.False(result);
     }
 }
